complete booking, search and paged

+++ src/main/java/vn/com/pn/repository/currencyunit/CurrencyUnitRepository.java
@@ -3,5 +3,5 @@ package vn.com.pn.repository.currencyunit;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.com.pn.domain.CurrencyUnit;
 
-public interface CurrencyUnitRepository extends JpaRepository <CurrencyUnit, Integer> {
+public interface CurrencyUnitRepository extends JpaRepository <CurrencyUnit, Long> {
 }
