+++ src/main/java/vn/com/pn/repository/datecannotbooking/DateCanNotBookingRepository.java
@@ -3,5 +3,5 @@ package vn.com.pn.repository.datecannotbooking;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.com.pn.domain.DateCanNotBooking;
 
-public interface DateCanNotBookingRepository extends JpaRepository<DateCanNotBooking, Integer> {
+public interface DateCanNotBookingRepository extends JpaRepository<DateCanNotBooking, Long> {
 }
