+++ src/main/java/vn/com/pn/repository/forgotpasswordcode/ForgotPasswordCodeRepository.java
@@ -3,5 +3,5 @@ package vn.com.pn.repository.forgotpasswordcode;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.com.pn.domain.ForgotPasswordCode;
 
-public interface ForgotPasswordCodeRepository extends JpaRepository <ForgotPasswordCode, Integer> {
+public interface ForgotPasswordCodeRepository extends JpaRepository <ForgotPasswordCode, Long> {
 }
