+++ src/main/java/vn/com/pn/repository/hostcancellationpolicy/HostCancellationPolicyRepository.java
@@ -3,5 +3,5 @@ package vn.com.pn.repository.hostcancellationpolicy;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.com.pn.domain.HostCancellationPolicy;
 
-public interface HostCancellationPolicyRepository extends JpaRepository<HostCancellationPolicy, Integer> {
+public interface HostCancellationPolicyRepository extends JpaRepository<HostCancellationPolicy, Long> {
 }
