+++ src/main/java/vn/com/pn/repository/hostcategory/HostCategoryRepository.java
@@ -3,5 +3,5 @@ package vn.com.pn.repository.hostcategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.com.pn.domain.HostCategory;
 
-public interface HostCategoryRepository extends JpaRepository <HostCategory, Integer > {
+public interface HostCategoryRepository extends JpaRepository <HostCategory, Long > {
 }
