+++ src/main/java/vn/com/pn/repository/hostcity/HostCityRepository.java
@@ -3,5 +3,5 @@ package vn.com.pn.repository.hostcity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.com.pn.domain.HostCity;
 
-public interface HostCityRepository extends JpaRepository <HostCity, Integer> {
+public interface HostCityRepository extends JpaRepository <HostCity, Long> {
 }
