+++ src/main/java/vn/com/pn/repository/hostdiscount/HostDiscountRepository.java
@@ -3,5 +3,5 @@ package vn.com.pn.repository.hostdiscount;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.com.pn.domain.HostDiscount;
 
-public interface HostDiscountRepository extends JpaRepository<HostDiscount, Integer> {
+public interface HostDiscountRepository extends JpaRepository<HostDiscount, Long> {
 }
