+++ src/main/java/vn/com/pn/repository/hostimage/HostImageRepository.java
@@ -3,5 +3,5 @@ package vn.com.pn.repository.hostimage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.com.pn.domain.HostImage;
 
-public interface HostImageRepository extends JpaRepository<HostImage, Integer> {
+public interface HostImageRepository extends JpaRepository<HostImage, Long> {
 }
