+++ src/main/java/vn/com/pn/repository/hostreview/HostReviewRepository.java
@@ -7,7 +7,7 @@ import vn.com.pn.domain.HostReview;
 import java.util.ArrayList;
 import java.util.List;
 
-public interface HostReviewRepository extends JpaRepository <HostReview, Integer> {
+public interface HostReviewRepository extends JpaRepository <HostReview, Long> {
     @Query(value="SELECT star_rating  FROM host_reviews WHERE host_id = ?1", nativeQuery=true)
     List<Integer> findStarRatingByHostId(int hostId);
 }
