+++ src/main/java/vn/com/pn/repository/hostroomtype/HostRoomTypeRepository.java
@@ -3,5 +3,5 @@ package vn.com.pn.repository.hostroomtype;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.com.pn.domain.HostRoomType;
 
-public interface HostRoomTypeRepository extends JpaRepository <HostRoomType, Integer> {
+public interface HostRoomTypeRepository extends JpaRepository <HostRoomType, Long> {
 }
