+++ src/main/java/vn/com/pn/repository/procedurecheckin/ProcedureCheckInRepository.java
@@ -3,5 +3,5 @@ package vn.com.pn.repository.procedurecheckin;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.com.pn.domain.ProcedureCheckIn;
 
-public interface ProcedureCheckInRepository extends JpaRepository<ProcedureCheckIn, Integer> {
+public interface ProcedureCheckInRepository extends JpaRepository<ProcedureCheckIn, Long> {
 }
