+++ src/main/java/vn/com/pn/repository/rule/RuleRepository.java
@@ -3,5 +3,5 @@ package vn.com.pn.repository.rule;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.com.pn.domain.Rule;
 
-public interface RuleRepository extends JpaRepository <Rule, Integer> {
+public interface RuleRepository extends JpaRepository <Rule, Long> {
 }
